+++ server/FastFuel/FastFuel/Features/Menus/DTOs/MenuRequestDto.cs
@@ -9,7 +9,7 @@ public record MenuRequestDto
     /// <summary>
     /// The price.
     /// </summary>
-    public required uint Price { get; init; }
+    public required double Price { get; init; }
     /// <summary>
     /// The description.
     /// </summary>
