+++ server/FastFuel/FastFuel/Features/Menus/DTOs/MenuResponseDto.cs
@@ -11,7 +11,7 @@ public record MenuResponseDto : IIdentifiable
     /// <summary>
     /// The price.
     /// </summary>
-    public required uint Price { get; init; }
+    public required double Price { get; init; }
     /// <summary>
     /// The description.
     /// </summary>
